refactor!: change main module name to alien_invasion and remove unused files
