+++ alien_invasion/sprite_manager.py
@@ -2,9 +2,9 @@ from dataclasses import dataclass, field
 
 import inject
 
-from src import settings
-from src.entities.sprites import Sprite
-from src.game_state import GameState
+from alien_invasion import settings
+from alien_invasion.entities.sprites import Sprite
+from alien_invasion.game_state import GameState
 
 
 @dataclass
